requirements.txt
playwright
openai
python-dotenv
flask
flask-cors
